requirements.txt
anyio==3.6.2
azure-core==1.26.4
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
isodate==0.6.1
pycparser==2.21
pydantic==1.10.7
pymssql==2.2.7
pyodbc==4.0.39
python-decouple==3.8
python-dotenv==1.0.0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
